FEAT: JWT 예외처리 적용

+++ src/main/java/com/team2/levelog/comment/service/CommentService.java
@@ -13,6 +13,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import static com.team2.levelog.global.GlobalResponse.code.ErrorCode.NO_ACCESS;
+
 @Service
 @RequiredArgsConstructor
 public class CommentService {
@@ -20,25 +22,11 @@ public class CommentService {
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
 
-    private final UserRepository userRepository; // 테스트를 위해 임시로 유저 끌어오기위한 의존성주입
-
-//    // 댓글, 대댓글 조회
-//    public List<Comment> getComment(Long postId){
-//        List<Comment> commentList = commentRepository.findAllByPost_Id(postId);
-//
-//
-//        return
-//    }
-
-
     // 댓글 작성
     public void createComment(Long postId, CommentRequestDto commentRequestDto, User user){
-//    public void createComment(Long postId, CommentRequestDto commentRequestDto){
         Post post = postRepository.findById(postId).orElseThrow(
                 ()-> new CustomException(ErrorCode.POST_NOT_FOUND)
         );
-
-//        User user = userRepository.findById(1L).orElse(null); // 테스트 유저
 
         Comment newComment = new Comment(post, commentRequestDto, user);
         commentRepository.save(newComment);
@@ -46,7 +34,6 @@ public class CommentService {
 
     // 대댓글 작성
     public void createReply(Long postId, Long commentsId, CommentRequestDto commentRequestDto, User user){
-//    public void createReply(Long postId, Long commentsId, CommentRequestDto commentRequestDto){
         Post post = postRepository.findById(postId).orElseThrow(
                 ()-> new CustomException(ErrorCode.POST_NOT_FOUND)
         );
@@ -54,8 +41,6 @@ public class CommentService {
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
                 ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
-
-//        User user = userRepository.findById(1L).orElse(null); // 테스트 유저
 
         Comment newComment = new Comment(post, comment, commentRequestDto, user);
         commentRepository.save(newComment);
@@ -64,20 +49,26 @@ public class CommentService {
     // 댓글, 대댓글 수정
     @Transactional
     public void modifyComment(Long commentsId, CommentRequestDto commentRequestDto, User user){
-//    public void modifyComment(Long commentsId, CommentRequestDto commentRequestDto){
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
                 ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
+
+        if(!user.getNickname().equals(comment.getNickname())){
+            throw new CustomException(ErrorCode.NO_ACCESS);
+        }
 
         comment.update(commentRequestDto);
     }
 
     // 댓글, 대댓글 삭제
     public void deleteComment(Long commentsId, User user){
-//    public void deleteComment(Long commentsId){
         Comment comment = commentRepository.findById(commentsId).orElseThrow(
                 ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND)
         );
+
+        if(!user.getNickname().equals(comment.getNickname())){
+            throw new CustomException(ErrorCode.NO_ACCESS);
+        }
 
         commentRepository.delete(comment);
     }
